cli > dev prompt styling & error handling

+++ lib/cli/commands/theme/dev.ts
@@ -1,28 +1,48 @@
 import { cwd } from 'process';
 import { readFileSync } from 'fs';
+import { clear } from 'console';
 import chokidar from 'chokidar';
 import kleur from 'kleur';
 import type { CLI, CommandDefinition } from '../types';
+import type { FileEventOptions } from './types';
 import stdout from '@/utils/system/stdout';
 import { getCurrentThemeId } from '@/utils/common';
 import config from '@/config';
 
-function logFileEvent(key: string, path: string) {
-  const tag = `[${kleur.bold().green(key)}]`;
-  return stdout.log(`${tag} ${path}`);
+const sizeFormatter = Intl.NumberFormat('en', {
+  notation: 'compact',
+  style: 'unit',
+  unit: 'byte',
+  unitDisplay: 'narrow',
+});
+
+const eventLogTagMap: { [key: string]: () => string } = {
+  error: () => kleur.bold().red('[error]'),
+  add: () => kleur.bold().green('[created]'),
+  change: () => kleur.bold().blue('[updated]'),
+  unlink: () => kleur.bold().yellow('[deleted]'),
+};
+
+function logFileEvent(options: FileEventOptions) {
+  const tag = eventLogTagMap[options.event]();
+
+  return stdout.log(`${tag} ${kleur.underline().white(options.path)} - ${sizeFormatter.format(options.size)} | ${options.roundtrip}ms \n`);
 }
 
 export default function command(cli: CLI): CommandDefinition {
   return {
     name: 'dev',
     group: 'theme',
-    description: 'stats a dev server and watches over the current directory',
+    description: 'starts a dev server and watches over the current directory',
 
     action: async () => {
       if (!cli.client.isAuthenticated())
         return stdout.error('You must be logged into a store to use this command.');
 
       const themeId = await getCurrentThemeId(cwd());
+
+      clear();
+      stdout.log('Watching theme files for changes.. \n');
 
       chokidar
         .watch(config.THEME_FILE_TYPES, {
@@ -32,30 +52,57 @@ export default function command(cli: CLI): CommandDefinition {
             stabilityThreshold: 50,
           },
         })
-        .on('all', (event, path) => {
-          const [filetype, filename] = path.split('/', 2);
+        .on('all', async (event, path, stats) => {
+          const start = new Date().getTime();
 
-          if (!config.THEME_FILE_TYPES.includes(filetype))
-            return;
+          try {
+            const [filetype, filename] = path.split('/', 2);
 
-          logFileEvent(event, path);
+            if (!config.THEME_FILE_TYPES.includes(filetype))
+              return;
 
-          switch (event) {
-            case 'add' || 'change':
-              cli.client.updateFile(themeId, {
-                file_type: filetype,
-                file_name: filename,
-                file_operation: 'save',
-                file_content: readFileSync(path, { encoding: 'utf-8', flag: 'r' }),
-              });
-              break;
-            case 'unlink':
-              cli.client.deleteFile(themeId, {
-                file_type: filetype,
-                file_name: filename,
-                file_operation: 'delete',
-              });
-              break;
+            if (!['add', 'change', 'unlink'].includes(event))
+              return;
+
+            switch (event) {
+              case 'add':
+              case 'change':
+                await cli.client.updateFile(themeId, {
+                  file_type: filetype,
+                  file_name: filename,
+                  file_operation: 'save',
+                  file_content: readFileSync(path, { encoding: 'utf-8', flag: 'r' }),
+                });
+
+                break;
+              case 'unlink':
+                await cli.client.deleteFile(themeId, {
+                  file_type: filetype,
+                  file_name: filename,
+                  file_operation: 'delete',
+                });
+
+                break;
+            }
+
+            logFileEvent({
+              path,
+              event,
+              size: stats!.size,
+              roundtrip: new Date().getTime() - start,
+            });
+          }
+
+          catch (err) {
+            logFileEvent({
+              path,
+              event: 'error',
+              size: stats!.size,
+              roundtrip: new Date().getTime() - start,
+            });
+
+            if (err instanceof Error)
+              stdout.info(`message: ${err.message}`);
           }
         });
     },
